src/scraper/scrape.py
import os
from dotenv import load_dotenv

# Logic for Telegram scraping for Task 1 will go here.

def scrape_telegram_data():
    load_dotenv()
    api_id = os.getenv('TELEGRAM_API_ID')
    api_hash = os.getenv('TELEGRAM_API_HASH')

    print("Starting Telegram scraping process...")
    print(f"Using API ID: {'*' * 5}{api_id[-4:] if api_id else 'Not Set'}") # Avoid printing full key
    
    # 1. Initialize Telethon client.
    # 2. Define the list of channels to scrape.
    # 3. Iterate through channels and fetch messages.
    # 4. Save messages and images to the partitioned data lake (data/raw/...).
    # 5. Implement logging.

    print("Telegram scraping complete.")


if __name__ == "__main__":
    scrape_telegram_data()